ATTEMPTS at making sign up work

+++ event/models.py
@@ -8,8 +8,8 @@ class Event(models.Model):
     event_start_date = models.DateTimeField("event start date")
     event_end_date = models.DateTimeField("event end date")
     event_add_date = models.DateTimeField("event add date")
-    event_price = models.FloatField()
+    event_price = models.PositiveIntegerField()
     event_description = models.TextField(max_length=1200)
-    
+
     def event_already_happened(self):
         return evemt_end_date >= timezone.now()
